+++ docker/core/mkmetadata/src/mk_lib_database_metadata_collection.rs
@@ -7,16 +7,16 @@ use serde::{Serialize, Deserialize};
 
 pub async fn mk_lib_database_metadata_collection_count(pool: &sqlx::PgPool,
                                                        search_value: String)
-                                                       -> Result<i32, sqlx::Error> {
+                                                       -> Result<i64, sqlx::Error> {
     if search_value != "" {
-        let row: (i32, ) = sqlx::query_as("select count(*) from mm_metadata_collection \
+        let row: (i64, ) = sqlx::query_as("select count(*) from mm_metadata_collection \
             where mm_metadata_collection_name = $1")
             .bind(search_value)
             .fetch_one(pool)
             .await?;
         Ok(row.0)
     } else {
-        let row: (i32, ) = sqlx::query_as("select count(*) from mm_metadata_collection")
+        let row: (i64, ) = sqlx::query_as("select count(*) from mm_metadata_collection")
             .fetch_one(pool)
             .await?;
         Ok(row.0)
