+++ docker/core/mktmdbnetfetchupdate/src/mk_lib_database_metadata_movie.rs
@@ -68,16 +68,16 @@ pub async fn mk_lib_database_metadata_movie_read(pool: &sqlx::PgPool,
 
 pub async fn mk_lib_database_metadata_movie_count(pool: &sqlx::PgPool,
                                                   search_value: String)
-                                                  -> Result<i32, sqlx::Error> {
+                                                  -> Result<i64, sqlx::Error> {
     if search_value != "" {
-        let row: (i32, ) = sqlx::query_as("select count(*) from mm_metadata_movie \
+        let row: (i64, ) = sqlx::query_as("select count(*) from mm_metadata_movie \
             where mm_metadata_name % $1")
             .bind(search_value)
             .fetch_one(pool)
             .await?;
         Ok(row.0)
     } else {
-        let row: (i32, ) = sqlx::query_as("select count(*) from mm_metadata_movie")
+        let row: (i64, ) = sqlx::query_as("select count(*) from mm_metadata_movie")
             .fetch_one(pool)
             .await?;
         Ok(row.0)
