+++ docker/core/mkwebapp/src/admin/bp_library.rs
@@ -16,9 +16,9 @@ struct TemplateAdminLibraryContext<> {
     pagination_bar: String,
 }
 
-#[get("/admin_library?<page>")]
+#[get("/admin_library/<page>")]
 pub async fn admin_library(sqlx_pool: &rocket::State<sqlx::PgPool>, user: AdminUser, page: i8) -> Template {
-    let total_pages: i32 = mk_lib_database_library::mk_lib_database_library_count(&sqlx_pool).await.unwrap() / 30;
+    let total_pages: i64 = mk_lib_database_library::mk_lib_database_library_count(&sqlx_pool).await.unwrap() / 30;
     let pagination_html = mk_lib_common_pagination::mk_lib_common_paginate(total_pages, page).await.unwrap();
     let library_list = mk_lib_database_library::mk_lib_database_library_read(&sqlx_pool, 0, 30).await.unwrap();
     Template::render("bss_admin/bss_admin_library", &TemplateAdminLibraryContext {
