+++ docker/core/mkwebapp/src/user/media/bp_media_collection.rs
@@ -16,9 +16,9 @@ struct TemplateMediaCollectionContext<> {
     pagination_bar: String,
 }
 
-#[get("/media/collection?<page>")]
+#[get("/media/collection/<page>")]
 pub async fn user_media_collection(sqlx_pool: &rocket::State<sqlx::PgPool>, user: User, page: i8) -> Template {
-    let total_pages: i32 = mk_lib_database_metadata_collection::mk_lib_database_metadata_collection_count(&sqlx_pool, String::new()).await.unwrap() / 30;
+    let total_pages: i64 = mk_lib_database_metadata_collection::mk_lib_database_metadata_collection_count(&sqlx_pool, String::new()).await.unwrap() / 30;
     let pagination_html = mk_lib_common_pagination::mk_lib_common_paginate(total_pages, page).await.unwrap();
     let collection_list = mk_lib_database_metadata_collection::mk_lib_database_metadata_collection_read(&sqlx_pool, String::new(), 0 ,30).await.unwrap();
     Template::render("bss_user/metadata/bss_user_metadata_movie_collection", &TemplateMediaCollectionContext {
