+++ docker/core/mkwebapp/src/user/media/bp_media_music_video.rs
@@ -16,9 +16,9 @@ struct TemplateMediaMusicVideoContext<> {
     pagination_bar: String,
 }
 
-#[get("/media/music_video?<page>")]
+#[get("/media/music_video/<page>")]
 pub async fn user_media_music_video(sqlx_pool: &rocket::State<sqlx::PgPool>, user: User, page: i8) -> Template {
-    let total_pages: i32 = mk_lib_database_media_music_video::mk_lib_database_media_music_video_count(&sqlx_pool, String::new()).await.unwrap() / 30;
+    let total_pages: i64 = mk_lib_database_media_music_video::mk_lib_database_media_music_video_count(&sqlx_pool, String::new()).await.unwrap() / 30;
     let pagination_html = mk_lib_common_pagination::mk_lib_common_paginate(total_pages, page).await.unwrap();
     let music_video_list = mk_lib_database_media_music_video::mk_lib_database_media_music_video_read(&sqlx_pool, String::new(), 0, 30).await.unwrap();
     Template::render("bss_user/media/bss_user_media_music_video", &TemplateMediaMusicVideoContext {
