+++ docker/core/mkwebapp/src/user/metadata/bp_meta_game.rs
@@ -16,9 +16,9 @@ struct TemplateMetaGameContext<> {
     pagination_bar: String,
 }
 
-#[get("/metadata/game?<page>")]
+#[get("/metadata/game/<page>")]
 pub async fn user_metadata_game(sqlx_pool: &rocket::State<sqlx::PgPool>, user: User, page: i8) -> Template {
-    let total_pages: i32 = mk_lib_database_metadata_game::mk_lib_database_metadata_game_count(&sqlx_pool, String::new()).await.unwrap() / 30;
+    let total_pages: i64 = mk_lib_database_metadata_game::mk_lib_database_metadata_game_count(&sqlx_pool, String::new()).await.unwrap() / 30;
     let pagination_html = mk_lib_common_pagination::mk_lib_common_paginate(total_pages, page).await.unwrap();
     let game_list = mk_lib_database_metadata_game::mk_lib_database_metadata_game_read(&sqlx_pool, String::new(), 0, 30).await.unwrap();
     Template::render("bss_user/metadata/bss_user_metadata_game", &TemplateMetaGameContext {
