+++ docker/core/mkwebapp/src/user/metadata/bp_meta_music.rs
@@ -16,9 +16,9 @@ struct TemplateMetaMusicContext<> {
     pagination_bar: String,
 }
 
-#[get("/metadata/music?<page>")]
+#[get("/metadata/music/<page>")]
 pub async fn user_metadata_music(sqlx_pool: &rocket::State<sqlx::PgPool>, user: User, page: i8) -> Template {
-    let total_pages: i32 = mk_lib_database_metadata_music::mk_lib_database_metadata_music_count(&sqlx_pool, String::new()).await.unwrap() / 30;
+    let total_pages: i64 = mk_lib_database_metadata_music::mk_lib_database_metadata_music_count(&sqlx_pool, String::new()).await.unwrap() / 30;
     let pagination_html = mk_lib_common_pagination::mk_lib_common_paginate(total_pages, page).await.unwrap();
     let music_list = mk_lib_database_metadata_music::mk_lib_database_metadata_music_read(&sqlx_pool, String::new(), 0, 30).await.unwrap();
     Template::render("bss_user/metadata/bss_user_metadata_music_album", &TemplateMetaMusicContext {
