+++ docker/core/mkwebapp/src/user/metadata/bp_meta_person.rs
@@ -16,9 +16,9 @@ struct TemplateMetaPersonContext<> {
     pagination_bar: String,
 }
 
-#[get("/metadata/person?<page>")]
+#[get("/metadata/person/<page>")]
 pub async fn user_metadata_person(sqlx_pool: &rocket::State<sqlx::PgPool>, user: User, page: i8) -> Template {
-    let total_pages: i32 = mk_lib_database_metadata_person::mk_lib_database_metadata_person_count(&sqlx_pool, String::new()).await.unwrap() / 30;
+    let total_pages: i64 = mk_lib_database_metadata_person::mk_lib_database_metadata_person_count(&sqlx_pool, String::new()).await.unwrap() / 30;
     let pagination_html = mk_lib_common_pagination::mk_lib_common_paginate(total_pages, page).await.unwrap();
     let person_list = mk_lib_database_metadata_person::mk_lib_database_metadata_person_read(&sqlx_pool, String::new(), 0, 30).await.unwrap();
     Template::render("bss_user/metadata/bss_user_metadata_person", &TemplateMetaPersonContext {
