+++ src/mk_lib_database/src/media/mk_lib_database_media_book.rs
@@ -47,9 +47,9 @@ pub async fn mk_lib_database_media_book_read(pool: &sqlx::PgPool,
 
 pub async fn mk_lib_database_media_book_count(pool: &sqlx::PgPool,
                                               search_value: String)
-                                              -> Result<i32, sqlx::Error> {
+                                              -> Result<i64, sqlx::Error> {
     if search_value != "" {
-        let row: (i32, ) = sqlx::query_as("select count(*) from mm_metadata_book, \
+        let row: (i64, ) = sqlx::query_as("select count(*) from mm_metadata_book, \
             mm_media where mm_media_metadata_guid = mm_metadata_book_guid \
             and mm_metadata_book_name % $1")
             .bind(search_value)
@@ -57,7 +57,7 @@ pub async fn mk_lib_database_media_book_count(pool: &sqlx::PgPool,
             .await?;
         Ok(row.0)
     } else {
-        let row: (i32, ) = sqlx::query_as("select count(*) from mm_metadata_book, \
+        let row: (i64, ) = sqlx::query_as("select count(*) from mm_metadata_book, \
             mm_media where mm_media_metadata_guid = mm_metadata_book_guid")
             .fetch_one(pool)
             .await?;
