+++ src/mk_lib_database/src/media/mk_lib_database_media_movie.rs
@@ -69,15 +69,15 @@ pub async fn mk_lib_database_media_movie_read(pool: &sqlx::PgPool,
 
 pub async fn mk_lib_database_media_movie_count(pool: &sqlx::PgPool,
                                                search_value: String)
-                                               -> Result<i32, sqlx::Error> {
+                                               -> Result<i64, sqlx::Error> {
     if search_value != "" {
-        let row: (i32, ) = sqlx::query_as("")
+        let row: (i64, ) = sqlx::query_as("")
             .bind(search_value)
             .fetch_one(pool)
             .await?;
         Ok(row.0)
     } else {
-        let row: (i32, ) = sqlx::query_as("")
+        let row: (i64, ) = sqlx::query_as("")
             .fetch_one(pool)
             .await?;
         Ok(row.0)
