+++ src/mk_lib_database/src/media/mk_lib_database_media_music_video.rs
@@ -36,9 +36,9 @@ pub async fn mk_lib_database_media_music_video_read(pool: &sqlx::PgPool,
 
 pub async fn mk_lib_database_media_music_video_count(pool: &sqlx::PgPool,
                                                      search_value: String)
-                                                     -> Result<i32, sqlx::Error> {
+                                                     -> Result<i64, sqlx::Error> {
     if search_value != "" {
-        let row: (i32, ) = sqlx::query_as("select count(*) from mm_metadata_music_video, mm_media \
+        let row: (i64, ) = sqlx::query_as("select count(*) from mm_metadata_music_video, mm_media \
             where mm_media_metadata_guid = mm_metadata_music_video_guid group \
             and mm_media_music_video_song % $1")
             .bind(search_value)
@@ -46,7 +46,7 @@ pub async fn mk_lib_database_media_music_video_count(pool: &sqlx::PgPool,
             .await?;
         Ok(row.0)
     } else {
-        let row: (i32, ) = sqlx::query_as("select count(*) from mm_metadata_music_video, mm_media \
+        let row: (i64, ) = sqlx::query_as("select count(*) from mm_metadata_music_video, mm_media \
             where mm_media_metadata_guid = mm_metadata_music_video_guid")
             .fetch_one(pool)
             .await?;
